revisi pak Fajar sebagian udah

+++ View/AdminProfil.php
@@ -37,7 +37,7 @@
           <td><?php echo "$row->email"; ?></td>
         </tr>
         <tr>
-          <td>Instansi &nbsp;&nbsp;</td>
+          <td>Program Studi &nbsp;&nbsp;</td>
           <td>:&nbsp;</td>
           <td><?php echo "$row->instansi"; ?></td>
         </tr>
@@ -59,7 +59,7 @@
       echo "    <div class='wrapper-action'>";
       echo "       <a href='" . $dataset->link_download . "'><i style='vertical-align: -25%;' class='fas fa-download fa-2x'></i></a>";
       echo "       <a href='AdminEditDataset.php?judul=$dataset->judul' class='btn btn-btn btn-success' role='button'>edit</a>";
-      echo "       <a href='../Config/DeleteDatasetAdmin.php?judul=$dataset->judul&id=$id' class='btn btn-danger' role='button'>hapus</a>";
+      echo "       <a href='../Config/DeleteDatasetAdmin.php?judul=$dataset->judul&id=$id' class='btn btn-danger' role='button' onclick=\"return confirm('Yakin Hapus?');\">hapus</a>";
       echo "     </div>";
       echo "</div>";
       echo "<br><br>";
